app-deploy/guestbook/app_code/requirements.txt
flask
mysql-connector-python
cryptography